seperate files into: cleverbot, conversations, rurubot.js
