+++ src/cleverbot.js
@@ -1,5 +1,5 @@
 // cleverbot instance (for fallback)
-let cleverbot = new(require("cleverbot.io"))(process.env.CLEVERBOT_API_USER, process.env.CLEVERBOT_API_KEY);
+const cleverbot = new(require("cleverbot.io"))(process.env.CLEVERBOT_API_USER, process.env.CLEVERBOT_API_KEY);
 cleverbot.setNick("Dry");
 cleverbot.create(function(err, session) {
     if (err) {
@@ -8,3 +8,5 @@ cleverbot.create(function(err, session) {
         console.log('cleverbot create success.');
     }
 });
+
+module.exports = cleverbot;
